Fixed Header, Nav Links and Slider Fixed
A header with fixed property has been fixed, navigation links are fixed however remapping is needed to the URI. The slider is fixed as this tends to get overlapped by the header and nav bar so a spacer div is used to overcome the problem. A bit of sliders image is oversized then it is supposed to be in the container-fluid div, need to fix and refine at the end.